requirements.txt
streamlit
langgraph
python-pptx
PyPDF2
python-dotenv
google-generativeai